Time: 0 ms (100.00%) | Memory: 111.4 MB (11.10%) - LeetSync

+++ 2432-number-of-zero-filled-subarrays/number-of-zero-filled-subarrays.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
         long long ans=0;
-        long long curr=0;
+        int curr=0;
         for(int i:nums){
             if(i==0){
                 curr++;
